+++ robotframework-ls/docs/changelog.md
@@ -6,6 +6,7 @@ Next
 - If a user is paused in a log message break he should still be able to evaluate keywords based on the previous stack context.
 - Right click on folder and selecting `Robot Framework -> Run/Debug Tests/Tasks Suite` properly runs tests again. [#776](https://github.com/robocorp/robotframework-lsp/issues/776)
 - In the `interactive console`, if a multi-line statement is entered without a `***` section prefix, properly indent considering new lines.
+- Restore behavior that showed errors for keyword callers in peek error window (configurable through `robot.run.peekError.showErrorsInCallers`).
 
 
 New in 1.3.0 (2022-10-06)